End of year project for java OOP
This was my first project as part of the HNC module in Java OOP in May 2017. I had to work through the different phases of the SDLC and create a object orientated solution for a Gas Company. I used UML Diagrams, Flowcharts, Pseudocode, Data Dictionaries, Units Test Program and a User guide.